requirements.txt
Package             Version
------------------- -----------
aiohappyeyeballs    2.6.1
aiohttp             3.12.7
aiosignal           1.3.2
altgraph            0.17.2
annotated-types     0.7.0
anyio               4.9.0
async-timeout       5.0.1
attrs               25.3.0
beautifulsoup4      4.13.4
black               25.1.0
certifi             2025.4.26
charset-normalizer  3.4.2
click               8.1.8
contourpy           1.3.0
cycler              0.12.1
datasets            3.6.0
dill                0.3.8
distlib             0.4.0
exceptiongroup      1.3.0
fastapi             0.115.12
filelock            3.18.0
flake8              7.2.0
fonttools           4.60.2
frozenlist          1.6.0
fsspec              2025.3.0
future              0.18.2
gdown               5.2.0
h11                 0.16.0
hf-xet              1.1.2
huggingface-hub     0.32.4
idna                3.10
importlib_resources 6.5.2
Jinja2              3.1.6
joblib              1.5.1
kagglehub           0.3.13
kiwisolver          1.4.7
macholib            1.15.2
markdown-it-py      3.0.0
MarkupSafe          3.0.2
matplotlib          3.9.4
mccabe              0.7.0
mdurl               0.1.2
mpmath              1.3.0
multidict           6.4.4
multiprocess        0.70.16
mypy_extensions     1.1.0
networkx            3.2.1
numpy               2.0.2
packaging           25.0
pandas              2.2.3
pathspec            0.12.1
pillow              11.3.0
pip                 25.3
platformdirs        4.3.8
propcache           0.3.1
pyarrow             20.0.0
pycodestyle         2.13.0
pydantic            2.11.5
pydantic_core       2.33.2
python-dotenv       1.0.0
pyflakes            3.3.2
Pygments            2.19.2
pyparsing           3.3.1
PySocks             1.7.1
python-dateutil     2.9.0.post0
pytz                2025.2
PyYAML              6.0.2
regex               2024.11.6
requests            2.32.3
rich                14.2.0
safetensors         0.5.3
scikit-learn        1.6.1
scipy               1.13.1
setuptools          58.0.4
six                 1.15.0
sniffio             1.3.1
soupsieve           2.7
starlette           0.46.2
sympy               1.14.0
threadpoolctl       3.6.0
tokenizers          0.21.1
tomli               2.2.1
torch               2.7.1
tqdm                4.67.1
transformers        4.52.4
typing_extensions   4.14.0
typing-inspection   0.4.1
tzdata              2025.2
urllib3             2.4.0
uvicorn             0.34.3
virtualenv          20.35.4
wheel               0.37.0
xxhash              3.5.0
yarl                1.20.0
zipp                3.23.0
